fail if rhsm subscribe fails
If the user supplies incorrect credentials or something goes wrong with
the RH Subsciption registation, the install should fail immediately.
Without this, none of the required services for the bastion and cluster
will be configured.

This was not failing before due to limitations with terraform's inline
and heredoc useage. This format seems to fail fast.

The other changes are to ensure that if the script is re-run and
directories are already moved (the workaround mentioned in the tf code),
that this does not fail the inline script.

Fixes https://github.com/ocp-power-automation/ocp4-upi-powervs/issues/75

Signed-off-by: Christy Norman <christy@linux.vnet.ibm.com>